+++ pl_examples/basic_examples/mnist_datamodule.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 import platform
 from typing import Optional
+from warnings import warn
 
 from torch.utils.data import DataLoader, random_split
 
@@ -55,8 +56,10 @@ class MNISTDataModule(LightningDataModule):
             normalize: If true applies image normalize
         """
         super().__init__(*args, **kwargs)
-        if platform.system() == "Windows":
-            # see: https://stackoverflow.com/a/59680818/4521646
+        if num_workers and platform.system() == "Windows":
+            # see: https://stackoverflow.com/a/59680818
+            warn(f"You have requested num_workers={num_workers} on Windows,"
+                 " but currently recommended is 0, so we set it for you")
             num_workers = 0
 
         self.dims = (1, 28, 28)